requirements.txt
transformers[torch]
gradio
cohere
python-dotenv